requirements.txt
python-telegram-bot
googletrans==4.0.0rc1
